+++ earthdaily/earthdatastore/cube_utils/__init__.py
@@ -1,9 +1,9 @@
 import logging
 from datetime import datetime
-
+from collections import defaultdict
+import pandas as pd
 import geopandas as gpd
 import numpy as np
-import pandas as pd
 import pytz
 import xarray as xr
 from rasterio.enums import Resampling
@@ -70,7 +70,7 @@ def _autofix_unfrozen_coords_dtype(ds):
     return ds
 
 
-def _cube_odc(items_collection, assets=None, times=None, dtype="float32", **kwargs):
+def _cube_odc(items_collection, assets=None, times=None, dtype="float32", properties=False, **kwargs):
     from odc import stac
 
     if "epsg" in kwargs:
@@ -91,9 +91,26 @@ def _cube_odc(items_collection, assets=None, times=None, dtype="float32", **kwar
         groupby=None,
         **kwargs,
     )
-
+    if properties:
+        metadata = defaultdict(list)
+        for i in items_collection:
+            if isinstance(properties,str):
+                metadata[properties].append(i.properties[properties])
+            else:
+                for k, v in i.properties.items():
+                    if isinstance(properties,list):
+                        if k not in properties:                    
+                            continue
+                    if isinstance(v, list):
+                        v = str(v)
+                    metadata[k].append(v)
+        # to avoid mismatch if some properties are not available on all items
+        df = pd.DataFrame.from_dict(metadata,orient='index').T
+        # convert to xarray needs
+        metadata = {k: ("time",v.tolist()) for k,v in df.items()}
+        # assign metadata as coords
+        ds = ds.assign_coords(**metadata)
     return ds
-
 
 def _cube_stackstac(items_collection, assets=None, times=None, **kwargs):
     from stackstac import stack
@@ -111,7 +128,6 @@ def _cube_stackstac(items_collection, assets=None, times=None, **kwargs):
         assets=assets,
         rescale=False,
         xy_coords="center",
-        properties=True,
         **kwargs,
     )
     ds = ds.to_dataset(dim="band")
@@ -135,6 +151,7 @@ def datacube(
     groupby_date="mean",
     common_band_names=True,
     cross_calibration_items: list | None = None,
+    properties:(bool | str | list) = False,
     **kwargs,
 ):
     logging.info(f"Building datacube with {len(items_collection)} items")
@@ -161,11 +178,12 @@ def datacube(
         kwargs["bounds_latlon"] = list(
             GeometryManager(intersects).to_geopandas().to_crs(epsg=4326).total_bounds
         )
-
+        
     ds = engines[engine](
         items_collection,
         assets=assets_keys if isinstance(assets, dict) else assets,
         times=times,
+        properties=properties,
         **kwargs,
     )
     nodatas = {}