+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.5] - 
 
+### Added
+
+- "properties" parameters for datacube creation.
+
 ## [0.0.4] - 2024-01-24
  
 ### Fixed